ransomeware/requirements.txt
altgraph==0.17.2
apturl==0.5.2
async-generator==1.10
attrs==21.4.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.0
chardet==4.0.0
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cryptography==36.0.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==2.1.2
future==0.18.2
h11==0.13.0
httplib2==0.14.0
idna==2.8
img2pdf==0.4.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
outcome==1.1.0
packaging==21.3
pandas==1.4.1
paramiko==2.6.0
pdfminer.six==20211012
pdfplumber==0.6.0
pexpect==4.6.0
pikepdf==5.1.1
Pillow==9.0.1
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.14.1
pycups==1.9.73
PyGObject==3.36.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.1.5
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
systemd-python==234
tk==0.1.0
tkintertable==1.3.3
trio==0.20.0
trio-websocket==0.9.2
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
Wand==0.6.7
Werkzeug==2.1.2
wsproto==1.1.0
xkit==0.0.0
zipp==3.8.0
